refactor: Payment 관련 패키지 구조 변경 및 인터페이스 위치 수정
PaymentRequest, PaymentResponse 등 Payment 관련 DTO를 application 패키지로 이동하여 계층 구조를 정리하고, PaymentClient 인터페이스를 application 패키지로 재배치했습니다. 이를 통해 모듈간 책임 분리를 명확히 했습니다.
