+++ src/main/java/roomescape/reservation/application/PaymentClient.java
@@ -1,0 +1,9 @@
+package roomescape.reservation.application;
+
+import roomescape.reservation.application.dto.PaymentRequest;
+import roomescape.reservation.application.dto.PaymentResponse;
+
+public interface PaymentClient {
+
+    PaymentResponse pay(PaymentRequest request);
+}
