--- src/main/java/roomescape/reservation/infrastructure/PaymentClient.java
@@ -1,9 +1,0 @@
-package roomescape.reservation.infrastructure;
-
-import roomescape.reservation.infrastructure.dto.PaymentRequest;
-import roomescape.reservation.infrastructure.dto.PaymentResponse;
-
-public interface PaymentClient {
-
-    PaymentResponse pay(PaymentRequest request);
-}
