+++ src/main/java/roomescape/reservation/infrastructure/dto/TossPaymentRequest.java
@@ -1,5 +1,7 @@
 package roomescape.reservation.infrastructure.dto;
 
+import roomescape.reservation.application.dto.PaymentRequest;
+
 public record TossPaymentRequest(
         Long amount,
         String orderId,
